Fix asymmetric encode/decode limits for max payload
Previously, MUNGE_MAXIMUM_REQ_LEN (1 MiB) was applied to both encode
and decode operations.  This created an asymmetry: payloads could
be encoded up to ~1 MiB, but decode was limited to ~767 KiB due to
base64 expansion (~33% overhead).

- Define explicit MUNGE_MAXIMUM_PAYLOAD_LEN (1 MiB) for payloads,
  and ensure credential payloads can utilize that full limit
  regardless of encoding options or metadata overhead.

- Set MUNGE_MAXIMUM_REQ_LEN to 2x the payload size to accommodate
  base64 encoding and metadata overhead.

- Enforce maximum limits in munge/unmunge clients, libmunge, and
  munged for defense-in-depth.

- Refactor read_data_from_file() to use chunked reading with overflow
  protection.  Also, fix an off-by-one error that prevented using
  the maximum payload.

- Add comprehensive tests including debug-mode bypass to validate
  limits in libmunge.

Users can now encode and decode payloads up to the full 1 MiB limit,
ensuring credentials round-trip successfully at the maximum size.

Tested:
- AlmaLinux 10.0 (x86-64-v2), 9.6, 8.10
- Arch Linux
- CentOS Stream 9
- Debian sid, 13.1, 12.12, 11.11, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5, 4
- Fedora 42, 41
- FreeBSD 14.3, 13.5
- macOS 15.6
- NetBSD 10.1, 9.4
- OpenBSD 7.7, 7.6
- openSUSE Leap 15.6, 15.5
- Ubuntu 25.04, 24.04.3, 22.04.5, 20.04.6, 18.04.6, 16.04.7

Closes #155

+++ src/libcommon/m_msg.h
@@ -34,8 +34,9 @@
 #endif /* HAVE_CONFIG_H */
 
 #include <inttypes.h>
-#include <munge.h>
 #include <netinet/in.h>                 /* for struct in_addr                */
+#include <stddef.h>
+#include <munge.h>
 
 
 /*****************************************************************************
@@ -130,9 +131,9 @@ void m_msg_reset (m_msg_t m);
 
 munge_err_t m_msg_bind (m_msg_t m, int sd);
 
-munge_err_t m_msg_send (m_msg_t m, m_msg_type_t type, int maxlen);
+munge_err_t m_msg_send (m_msg_t m, m_msg_type_t type, size_t maxlen);
 
-munge_err_t m_msg_recv (m_msg_t m, m_msg_type_t type, int maxlen);
+munge_err_t m_msg_recv (m_msg_t m, m_msg_type_t type, size_t maxlen);
 
 int m_msg_set_err (m_msg_t m, munge_err_t e, char *s);
